backend/score_project/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==25.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
dj-database-url
psycopg2-binary